drivers: dw1000: expose phy config struct
phy config struct can be used to define phy
configs in another place.
